Add authentication for get loan by id and get loans by account id

+++ server/src/loan/loan.service.ts
@@ -4,11 +4,13 @@ import { StatusCodes, ReasonPhrases } from 'http-status-codes';
 import { Loan } from './loan.entity';
 import { AppDataSource } from '../data-source';
 import AppError from '../utils/appError';
+import { Account } from '../account/account.entity';
+import { Roles } from '../enums/roles.enum';
 
 @injectable()
 export class LoanService {
   private loanRepo: Repository<Loan>;
-
+  private accountRepo: Repository<Account>;
   constructor() {
     this.loanRepo = AppDataSource.getRepository(Loan);
   }
@@ -20,13 +22,16 @@ export class LoanService {
     return loans;
   }
 
-  async getLoanById(id: string): Promise<Loan> {
+  async getLoanById(id: string, ownerId: string, role: Roles): Promise<Loan> {
     const loan = await this.loanRepo.findOne({ where: { id } });
     if (!loan) throw new AppError('Loan not found', StatusCodes.NOT_FOUND, ReasonPhrases.NOT_FOUND);
+    const account = await this.accountRepo.findOneBy({ id: loan.accountId });
+    if (ownerId !== account?.userId && role !== Roles.ADMIN)
+      throw new AppError('Access is denied', StatusCodes.FORBIDDEN, ReasonPhrases.FORBIDDEN);
     return loan;
   }
 
-  async getLoansByAccountId(accountId: string): Promise<Loan[]> {
+  async getLoansByAccountId(accountId: string, ownerId: string, role: Roles): Promise<Loan[]> {
     const loans = await this.loanRepo.find({ where: { accountId } });
     if (!loans.length)
       throw new AppError(
@@ -34,17 +39,22 @@ export class LoanService {
         StatusCodes.NOT_FOUND,
         ReasonPhrases.NOT_FOUND,
       );
+    const account = await this.accountRepo.findOneBy({ id: accountId });
+    if (account?.userId !== ownerId && role !== Roles.ADMIN)
+      throw new AppError('Access is denied', StatusCodes.FORBIDDEN, ReasonPhrases.FORBIDDEN);
     return loans;
   }
 
   async deleteLoan(id: string): Promise<string> {
-    const loan = await this.getLoanById(id);
+    const loan = await this.loanRepo.findOneBy({ id });
+    if (!loan) throw new AppError('Loan not found', StatusCodes.NOT_FOUND, ReasonPhrases.NOT_FOUND);
     await this.loanRepo.remove(loan);
     return 'Loan deleted successfully';
   }
 
   async updateLoanPaymentStatus(id: string, isPaid: boolean): Promise<string> {
-    const loan = await this.getLoanById(id);
+    const loan = await this.loanRepo.findOneBy({ id });
+    if (!loan) throw new AppError('Loan not found', StatusCodes.NOT_FOUND, ReasonPhrases.NOT_FOUND);
     loan.isPaid = isPaid;
     await this.loanRepo.save(loan);
     return `Loan now is ${isPaid ? 'paid' : 'not paid'}`;
